requirements.txt
rasterio
rasterstats
xarray
h5py
netcdf4
eccodes
cfgrib
python-dateutil
numpy
pandas
requests
pyshp
affine
pillow